SDKNEW-1747/feature - update folder name
