+++ src/main/java/com/perimeterx/api/PerimeterX.java
@@ -28,9 +28,9 @@ package com.perimeterx.api;
 import com.google.gson.Gson;
 import com.perimeterx.api.activities.ActivityHandler;
 import com.perimeterx.api.activities.BufferedActivityHandler;
-import com.perimeterx.api.additionals2s.credentialsIntelligence.AdditionalContext;
-import com.perimeterx.api.additionals2s.credentialsIntelligence.loginresponse.LoginResponseValidator;
-import com.perimeterx.api.additionals2s.credentialsIntelligence.loginresponse.LoginResponseValidatorFactory;
+import com.perimeterx.api.additionalContext.AdditionalContext;
+import com.perimeterx.api.additionalContext.credentialsIntelligence.loginresponse.LoginResponseValidator;
+import com.perimeterx.api.additionalContext.credentialsIntelligence.loginresponse.LoginResponseValidatorFactory;
 import com.perimeterx.api.providers.CombinedIPProvider;
 import com.perimeterx.api.providers.DefaultHostnameProvider;
 import com.perimeterx.api.providers.HostnameProvider;
