+++ src/main/java/com/perimeterx/api/additionalContext/credentialsIntelligence/loginrequest/RequestBodyExtractor.java
@@ -1,4 +1,4 @@
-package com.perimeterx.api.additionals2s.credentialsIntelligence.loginrequest;
+package com.perimeterx.api.additionalContext.credentialsIntelligence.loginrequest;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
