+++ src/main/java/com/perimeterx/api/additionalContext/credentialsIntelligence/loginresponse/LoginResponseValidator.java
@@ -1,4 +1,4 @@
-package com.perimeterx.api.additionals2s.credentialsIntelligence.loginresponse;
+package com.perimeterx.api.additionalContext.credentialsIntelligence.loginresponse;
 
 import com.perimeterx.http.ResponseWrapper;
