+++ src/main/java/com/perimeterx/api/additionalContext/credentialsIntelligence/protocol/CredentialsIntelligenceProtocol.java
@@ -1,6 +1,6 @@
-package com.perimeterx.api.additionals2s.credentialsIntelligence.protocol;
+package com.perimeterx.api.additionalContext.credentialsIntelligence.protocol;
 
-import com.perimeterx.api.additionals2s.credentialsIntelligence.UserLoginData;
+import com.perimeterx.api.additionalContext.credentialsIntelligence.UserLoginData;
 import com.perimeterx.models.configuration.credentialsIntelligenceconfig.LoginCredentials;
 
 public interface CredentialsIntelligenceProtocol {
