+++ src/main/java/com/perimeterx/api/additionalContext/credentialsIntelligence/protocol/V1CIProtocol.java
@@ -1,7 +1,7 @@
-package com.perimeterx.api.additionals2s.credentialsIntelligence.protocol;
+package com.perimeterx.api.additionalContext.credentialsIntelligence.protocol;
 
-import com.perimeterx.api.additionals2s.credentialsIntelligence.CIProtocol;
-import com.perimeterx.api.additionals2s.credentialsIntelligence.UserLoginData;
+import com.perimeterx.api.additionalContext.credentialsIntelligence.CIProtocol;
+import com.perimeterx.api.additionalContext.credentialsIntelligence.UserLoginData;
 import com.perimeterx.models.configuration.credentialsIntelligenceconfig.LoginCredentials;
 
 import static com.perimeterx.utils.HashAlgorithm.SHA256;
